snippet.py: bugs/noautowrap/nomake, colr.h: docs

+++ tools/snippet.py
@@ -123,13 +123,13 @@ USAGESTR = f"""{VERSIONSTR}
         {SCRIPT} [-D] [-n] [-q] [-w] -V
         {SCRIPT} [-D] [-n] [-q] [-d | -m | -r exe] -b
         {SCRIPT} [-D] [-n] [-q] [-d | -m | -r exe | -s] [-t name] -x [PATTERN] [-- ARGS...]
-        {SCRIPT} [-D] [-n] [-q] [-d | -m | -r exe | -s] [-t name] [CODE] [-- ARGS...]
-        {SCRIPT} [-D] [-n] [-q] [-d | -m | -r exe | -s] [-t name] [-f file...] [-- ARGS...]
-        {SCRIPT} [-D] [-n] [-q] [-d | -m | -r exe | -s] [-t name] [-w] (-e [CODE] | -l) [-- ARGS...]
-        {SCRIPT} [-D] -E [-t name] -x [PATTERN] [-- ARGS...]
-        {SCRIPT} [-D] -E [-t name] [CODE] [-- ARGS...]
-        {SCRIPT} [-D] -E [-t name] [-f file...] [-- ARGS...]
-        {SCRIPT} [-D] -E [-t name] [-w] (-e [CODE] | -l) [-- ARGS...]
+        {SCRIPT} [-D] [-W] [-n] [-q] [-d | -m | -r exe | -s] [-M | -t name] [CODE] [-- ARGS...]
+        {SCRIPT} [-D] [-W] [-n] [-q] [-d | -m | -r exe | -s] [-M | -t name] [-f file...] [-- ARGS...]
+        {SCRIPT} [-D] [-W] [-n] [-q] [-d | -m | -r exe | -s] [-M | -t name] [-w] (-e [CODE] | -l) [-- ARGS...]
+        {SCRIPT} [-D] -E [-M | -t name] -x [PATTERN] [-- ARGS...]
+        {SCRIPT} [-D] [-W] -E [-M | -t name] [CODE] [-- ARGS...]
+        {SCRIPT} [-D] [-W] -E [-M | -t name] [-f file...] [-- ARGS...]
+        {SCRIPT} [-D] [-W] -E [-M | -t name] [-w] (-e [CODE] | -l) [-- ARGS...]
 
     Options:
         ARGS                   : Extra arguments for the compiler.
@@ -157,6 +157,10 @@ USAGESTR = f"""{VERSIONSTR}
         -h,--help              : Show this help message.
         -L,--listexamples      : List example code snippets in the source.
         -l,--last              : Re-run the last snippet.
+        -M,--nomake            : Do not run `make` to get compiler args.
+                                 Running make could cause side-effects if this
+                                 tool is executed from another project directory
+                                 (with a Makefile in it).
         -m,--memcheck          : Run the snippet through `valgrind`.
         -N,--listnames         : List example snippet names from the source.
         -n,--name              : Print the resulting binary name, for further
@@ -168,6 +172,11 @@ USAGESTR = f"""{VERSIONSTR}
         -t name,--target name  : Make target to get compiler flags from.
         -V,--viewlast          : View the last snippet that was compiled.
         -v,--version           : Show version.
+        -W,--noautowrap        : Don't use any auto includes or macros.
+                                 ColrC itself will not be compiled in, and it's
+                                 header won't be available. This is good for
+                                 simple C snippets, but not testing/developing
+                                 ColrC.
         -w,--wrapped           : Use the "wrapped" version, which is the resulting
                                  `.c` file for snippets.
         -x,--examples          : Use source examples as the snippets.
@@ -192,6 +201,9 @@ def main(argd):
     """ Main entry point, expects docopt arg dict as argd. """
     global status
     last_snippet_read()
+    if argd['--noautowrap']:
+        Snippet.do_wrap = False
+
     if argd['--quiet']:
         status = noop
     if argd['--sanitize']:
@@ -257,6 +269,11 @@ def main(argd):
         ]
         if not snippets:
             raise InvalidArg('no files read.')
+    elif argd['CODE'] and os.path.exists(argd['CODE']):
+        snippet = read_file(argd['CODE'])
+        if not snippet:
+            raise InvalidArg(f'cannot read file: argd["CODE"]')
+        snippets = [snippet]
 
     if argd['--editlast']:
         if argd['--wrapped']:
@@ -295,7 +312,7 @@ def main(argd):
         disasm=argd['--disasm'],
         memcheck=argd['--memcheck'],
         quiet=argd['--quiet'],
-        make_target=argd['--target'],
+        make_target=Disabled if argd['--nomake'] else argd['--target'],
     )
 
 
@@ -555,14 +572,19 @@ def get_gcc_cmd(
         cmd.extend(('-o', output_file))
     cmd.append(f'-iquote{COLR_DIR}')
 
-    try:
-        compiler, make_flags = get_make_flags(
-            user_args=[make_target] if make_target else None
-        )
-    except ValueError:
-        # No make flags available.
+    if make_target is Disabled:
+        debug('`make` flags were disabled.')
         compiler = None
         make_flags = []
+    else:
+        try:
+            compiler, make_flags = get_make_flags(
+                user_args=[make_target] if make_target else None
+            )
+        except ValueError:
+            # No make flags available.
+            compiler = None
+            make_flags = []
 
     if not compiler:
         compiler = 'gcc'
@@ -1031,6 +1053,7 @@ def run_snippets(
         disasm=False, memcheck=False, quiet=False, make_target=None):
     """ Compile and run several c code snippets. """
     errs = 0
+    sniplen = len(snippets)
     for snippet in snippets:
         binaryname = snippet.compile(
             user_args=compiler_args,
@@ -1047,7 +1070,11 @@ def run_snippets(
                 memcheck=memcheck,
                 quiet=quiet,
             )
-            errs += 1 if procresult['returncode'] else 0
+            if procresult['returncode']:
+                debug(f'Snippet process returned: {procresult["returncode"]}')
+                # Use the actual process return code for single runs.
+                errs += procresult['returncode'] if (sniplen == 1) else 1
+
             status_runtime(procresult['duration'])
     return errs
 
@@ -1172,6 +1199,20 @@ class CompileError(ValueError):
         rsn = f', {self.reason}' if self.reason else ''
         fpath = self.filepath or '<no filepath>'
         return f'Can\'t compile snippet{rsn}: {fpath}'
+
+
+class _Disabled(object):
+    """ Used to check whether a feature should be disabled (see uses of it).
+    """
+    def __repr__(self):
+        return 'Disabled()'
+
+    def __str__(self):
+        return '<disabled>'
+
+
+# A singleton instance, so `is Disabled` works.
+Disabled = _Disabled()
 
 
 class EditError(CompileError):
@@ -1410,6 +1451,7 @@ class Snippet(object):
     indent = 4
     code_indent = 4
     quiet_mode = False
+    do_wrap = True
 
     def __init__(self, code, name=None):
         self.code = code
@@ -1467,9 +1509,11 @@ class Snippet(object):
         last_snippet_write(self.code)
         basename = os.path.split(os.path.splitext(self.src_file)[0])[-1]
         objname = f'{basename}.o'
-        cfiles = [self.src_file, COLRC_FILE]
+        cfiles = [self.src_file]
+        if self.do_wrap:
+            cfiles.append(COLRC_FILE)
         colrobj = 'colr.o'
-        if os.path.exists(colrobj):
+        if self.do_wrap and os.path.exists(colrobj):
             try:
                 os.remove(colrobj)
                 debug(f'Removed existing: {colrobj}')
@@ -1493,7 +1537,8 @@ class Snippet(object):
         debug(f'Found object files: {", ".join(tmpobjnames)}')
         objnames = []
         objnames.extend(tmpobjnames)
-        objnames.append(colrobj)
+        if self.do_wrap:
+            objnames.append(colrobj)
         binaryname = temp_file_name(ext='.binary')
         linkcmd = get_gcc_cmd(
             objnames,
@@ -1595,6 +1640,9 @@ class Snippet(object):
             If main() is already defined, the snippet is not wrapped.
             If colr.h is already included, no duplicate include is added.
         """
+        if not self.do_wrap:
+            debug('Not doing auto includes/macros.')
+            return code
         line_table = set(line.strip() for line in code.splitlines())
         lines = []
         for defname in sorted(config['defines']):
